docs: adjust KDF-counter command docs
This commit adjusts the README.md documentation for the KDF-counter
command to match the implementation.

Prior to this the kdf struct in subprocess that dispatches KDF-counter
command invocations had a few divergences from the docs:

* If the test case has the "Deferred" property set to true, then the key
  argument provided to the wrapper is empty.
* The wrapper is expected to output three values: the input key (since
  for deferred tests it was generated module-side), the fixed counter
  data, and the derived key.

For deferred tests the returned key is written to the response `KeyIn`.
For non-deferred tests the returned key is verified to match the one
that was sent to the submodule as a command arg.