subprocess: generalize hkdf KDA
Previously only the KDA HKDF mode was implemented. In preparation for
supporting KDA OneStep No Counter we need to generalize the handler as
both are advertised as "KDA" algorithm and must be multiplexed on the
"mode".

Unfortunately there's little overlap between HKDF and OneStepNoCounter
test vectors, so we more or less want separate Process machinery after
dispatching on the mode.